+++ packages/taro-rn-supporter/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tarojs/rn-supporter",
-  "version": "3.2.0-canary.5",
+  "version": "3.2.0-canary.7",
   "description": "Taro rn supporter",
   "main": "dist/index.js",
   "scripts": {
@@ -31,8 +31,8 @@
     "npm": ">=6.0.0"
   },
   "dependencies": {
-    "@tarojs/helper": "3.2.0-canary.5",
-    "@tarojs/rn-style-transformer": "3.2.0-canary.5",
+    "@tarojs/helper": "3.2.0-canary.7",
+    "@tarojs/rn-style-transformer": "3.2.0-canary.7",
     "lodash": "^4.17.20",
     "metro": "0.58.0",
     "metro-react-native-babel-transformer": "^0.58.0",
