+++ packages/taro-rn-supporter/src/index.ts
@@ -1,5 +1,3 @@
 import { Supporter } from './Support'
-import { transform } from './taroTransformer'
-export { Supporter, transform }
-export default Supporter
+export default { Supporter }
 module.exports = Supporter
